Optimize updating variant name
